refactor(cli): extract migrate-hooks command
