+++ crates/git-smee-cli/src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod init;
 pub(crate) mod install;
+pub(crate) mod migrate_hooks;
 pub(crate) mod run;
